requirements.txt
schoolopy
pytextnow
pyyaml
dataclasses
